Library Reference

Petri.Petri — Module

Petri

Provides a modeling framework for representing and solving stochastic petri nets

Catlab.Graphics.Graphviz.Graph — Method

Graph(model::Model)

convert a Model into a GraphViz Graph. Transition are green boxes and states are blue circles. Arrows go from the input states to the output states for each transition.

Petri.Model — Type

Model{S,D}

Structure for representing the petri net model

represented by states and transition functions

Petri.Problem — Type

Problem{M<:Model, S, N}

Structure for representing a petri net problem

represented by a petri net model, initial state, and number of steps

Petri.NullPetri — Method

NullPetri(n::Int)

create a Petri net of $n$ states with no transitions

Petri.solve — Method

solve(p::Problem)

Evaluate petri net problem and return the final state

Petri.vectorfields — Method

vectorfields(m::Model)

Convert a petri model into a differential equation function that can be passed into DifferentialEquation.jl or OrdinaryDiffEq.jl solvers