Library Reference

Petri.Petri — Module

Petri

Provides a modeling framework for representing and solving stochastic petri nets

Petri.NullPetri — Constant

NullPetri

create a Petri net of no states and no transitions

Catlab.Graphics.Graphviz.Graph — Method

Graph(model::Model)

convert a Model into a GraphViz Graph. Transition are green boxes and states are blue circles. Arrows go from the input states to the output states for each transition.

DiffEqBase.ODEProblem — Method

ODEProblem(m::Model, u0, tspan, β)

Generate an OrdinaryDiffEq ODEProblem

DiffEqBase.SDEProblem — Method

SDEProblem(m::Model, u0, tspan, β)

Generate an StochasticDiffEq SDEProblem and an appropriate CallbackSet

Petri.Model — Type

Model{S,D}

Structure for representing the petri net model

represented by states and transition functions

Petri.EmptyPetri — Method

EmptyPetri(n::Int)

create a Petri net of $n$ states with no transitions

Petri.EmptyPetri — Method

EmptyPetri(s::Vector{T})

create a Petri net with states $s$ with no transitions

Petri.vectorfield — Method

vectorfield(m::Model)

Convert a petri model into a differential equation function that can be passed into DifferentialEquation.jl or OrdinaryDiffEq.jl solvers